Style tom app
Add AdvertisementCard
Add routes for create, list of demands
apply styles for dackl tabs
Add customt dackl tablist
Apply auto-load by using watchers
