+++ apps/tom/src/components/tabs/TabItemType.ts
@@ -1,0 +1,4 @@
+export type TabItemType = {
+    id: string;
+    label: string;
+};
