+++ apps/tom/src/constants/solid-urls.ts
@@ -1,0 +1,6 @@
+export const bank = "https://bank.solid.aifb.kit.edu/profile/card#me";
+export const tax = "https://tax.solid.aifb.kit.edu/profile/card#me";
+export const documentCreationDemandShapeTreeUri = "https://solid.aifb.kit.edu/shapes/mandat/document.tree#documentCreationDemandTree";
+export const creditDemandShapeTreeUri = "https://solid.aifb.kit.edu/shapes/mandat/credit.tree#creditDemandTree";
+export const orderShapeTreeUri = "https://solid.aifb.kit.edu/shapes/mandat/credit.tree#creditOrderTree";
+export const orderContainer = "https://bank.solid.aifb.kit.edu/credits/orders/";
