+++ apps/tom/src/types/Advertisement.ts
@@ -1,0 +1,13 @@
+
+export interface Advertisement {
+    id: string;
+    validUntil: string;
+    inbox: string;
+    comment: string;
+    label: string;
+    lowestInterestRate: string;
+    minCreditPeriodMonths: string;
+    maxCreditPeriodMonths: string;
+    createdBy: string;
+    creatorIconURI: string;
+}
