+++ apps/tom/src/types/Demand.ts
@@ -1,0 +1,14 @@
+import {Offer} from "@/types/Offer";
+import {Order} from "@/types/Order";
+
+export interface Demand {
+    providerName: string;
+    providerWebID: string;
+    amount: number;
+    currency: string;
+    offer?: Offer;
+    order?: Order;
+    hasAccessRequest: string;
+    isAccessRequestGranted: string;
+    documentCreationDemand: string;
+}
