+++ apps/tom/src/types/Offer.ts
@@ -1,0 +1,6 @@
+
+export interface Offer {
+    id: string;
+    interestRate: number;
+    duration: string;
+}
