+++ apps/tom/src/types/Order.ts
@@ -1,0 +1,5 @@
+
+export interface Order {
+    id: string;
+    isTerminated: boolean;
+}
