+++ apps/tom/src/views/Services.vue
@@ -1,0 +1,11 @@
+<script setup lang="ts">
+
+</script>
+
+<template>
+<div>Services </div>
+</template>
+
+<style scoped>
+
+</style>
