+++ libs/theme/src/datev-theme/_extensions.scss
@@ -66,3 +66,7 @@ a:not(.p-accordion-header-link) {
 .break-keep {
     word-break: keep-all;
 }
+
+.bg-gradient-blue {
+    background: linear-gradient(to right, #195B78, #287F8F);
+}
