+++ libs/utils/index.ts
@@ -1,3 +1,5 @@
 export * from './src/setupApp'
 export * from './src/requireStyles'
 export * from './src/wait'
+export * from './src/fetchStoreOf'
+export * from './src/getContainerUris'
