+++ libs/utils/src/getContainerUris.ts
@@ -1,0 +1,9 @@
+import {getDataRegistrationContainers, Session} from "@shared/solid";
+
+export async function getContainerUris(webId: string, shapeTreeUri: string, session: Session): Promise<string[]> {
+    return await getDataRegistrationContainers(
+        webId,
+        shapeTreeUri,
+        session
+    );
+}
